fix(lint): extract rehydrate_manifests to reduce replication.service line count
replication.service.ts was 303 counted lines (3 over the 300-line max-lines
limit). Extracted the rehydrate_manifests private method into a standalone
run_manifest_rehydration function in rehydration-manifests-runner.ts, reducing
the service file to 274 lines.

Made-with: Cursor
